Initial split into collections module
