+++ Cargo.toml
@@ -24,6 +24,7 @@ members = [
   "circuit",
   "circuit/account",
   "circuit/algorithms",
+  "circuit/collections",
   "circuit/environment",
   "circuit/environment/witness",
   "circuit/network",
@@ -39,6 +40,7 @@ members = [
   "console",
   "console/account",
   "console/algorithms",
+  "console/collections",
   "console/network",
   "console/program",
   "console/types",
