+++ circuit/collections/Cargo.toml
@@ -1,0 +1,43 @@
+[package]
+name = "snarkvm-circuit-collections"
+version = "0.7.5"
+authors = [ "The Aleo Team <hello@aleo.org>" ]
+description = "Collections circuit library for a decentralized virtual machine"
+license = "GPL-3.0"
+edition = "2021"
+
+[dependencies.console]
+package = "snarkvm-console-collections"
+path = "../../console/collections"
+version = "0.7.5"
+optional = true
+
+[dependencies.snarkvm-circuit-network]
+path = "../network"
+version = "0.7.5"
+
+[dependencies.snarkvm-circuit-types]
+path = "../types"
+version = "0.7.5"
+
+[dependencies.snarkvm-curves]
+path = "../../curves"
+version = "0.7.5"
+default-features = false
+
+[dependencies.snarkvm-fields]
+path = "../../fields"
+version = "0.7.5"
+default-features = false
+
+[dev-dependencies.snarkvm-utilities]
+path = "../../utilities"
+version = "0.7.5"
+
+[dev-dependencies.anyhow]
+version = "1.0.57"
+
+[features]
+default = ["enable_console"]
+# Use #[cfg(console)] instead.
+enable_console = ["console"]
