+++ circuit/collections/build.rs
@@ -14,20 +14,8 @@
 // You should have received a copy of the GNU General Public License
 // along with the snarkVM library. If not, see <https://www.gnu.org/licenses/>.
 
-#![forbid(unsafe_code)]
-#![allow(clippy::too_many_arguments)]
-
-pub mod bhp;
-pub use bhp::*;
-
-pub mod elligator2;
-pub use elligator2::Elligator2;
-
-pub mod pedersen;
-pub use pedersen::*;
-
-pub mod poseidon;
-pub use poseidon::*;
-
-pub mod traits;
-pub use traits::*;
+fn main() {
+    if cfg!(feature = "enable_console") {
+        println!("cargo:rustc-cfg=console");
+    }
+}
