+++ circuit/collections/src/merkle_tree/to_root.rs
