+++ circuit/program/README.md
@@ -1,5 +1,5 @@
-# snarkvm-circuit-aleo
+# snarkvm-circuit-program
 
-[![Crates.io](https://img.shields.io/crates/v/snarkvm-circuit-aleo.svg?color=neon)](https://crates.io/crates/snarkvm-circuit-aleo)
+[![Crates.io](https://img.shields.io/crates/v/snarkvm-circuit-program.svg?color=neon)](https://crates.io/crates/snarkvm-circuit-program)
 [![Authors](https://img.shields.io/badge/authors-Aleo-orange.svg)](https://aleo.org)
 [![License](https://img.shields.io/badge/License-GPLv3-blue.svg)](./LICENSE.md)
