+++ console/collections/Cargo.toml
@@ -1,0 +1,51 @@
+[package]
+name = "snarkvm-console-collections"
+version = "0.7.5"
+authors = [ "The Aleo Team <hello@aleo.org>" ]
+description = "Collections for a decentralized virtual machine"
+license = "GPL-3.0"
+edition = "2021"
+
+[dependencies.snarkvm-console-algorithms]
+path = "../algorithms"
+version = "0.7.5"
+
+[dependencies.snarkvm-console-network]
+path = "../network"
+version = "0.7.5"
+
+[dependencies.snarkvm-fields]
+path = "../../fields"
+version = "0.7.5"
+default-features = false
+
+[dependencies.snarkvm-utilities]
+path = "../../utilities"
+version = "0.7.5"
+
+[dependencies.anyhow]
+version = "1.0.57"
+
+[dependencies.itertools]
+version = "0.10.1"
+
+[dependencies.once_cell]
+version = "1.10.0"
+
+[dependencies.rayon]
+version = "1"
+optional = true
+
+[dependencies.serde]
+version = "1.0"
+features = ["derive"]
+
+[dev-dependencies.bincode]
+version = "1.3"
+
+[dev-dependencies.rand]
+version = "0.8"
+
+[features]
+default = ["parallel"]
+parallel = ["rayon"]
