+++ console/collections/README.md
@@ -1,0 +1,5 @@
+# snarkvm-console-collections
+
+[![Crates.io](https://img.shields.io/crates/v/snarkvm-console-collections.svg?color=neon)](https://crates.io/crates/snarkvm-console-collections)
+[![Authors](https://img.shields.io/badge/authors-Aleo-orange.svg)](https://aleo.org)
+[![License](https://img.shields.io/badge/License-GPLv3-blue.svg)](./LICENSE.md)
