+++ console/collections/src/merkle_tree/helpers/leaf_hash.rs
