+++ console/collections/src/merkle_tree/helpers/path_hash.rs
