+++ console/program/README.md
@@ -1,5 +1,5 @@
-# snarkvm-console-aleo
+# snarkvm-console-program
 
-[![Crates.io](https://img.shields.io/crates/v/snarkvm-console-aleo.svg?color=neon)](https://crates.io/crates/snarkvm-console-aleo)
+[![Crates.io](https://img.shields.io/crates/v/snarkvm-console-program.svg?color=neon)](https://crates.io/crates/snarkvm-console-program)
 [![Authors](https://img.shields.io/badge/authors-Aleo-orange.svg)](https://aleo.org)
 [![License](https://img.shields.io/badge/License-GPLv3-blue.svg)](./LICENSE.md)
